修复文件编码的BUG

+++ MatrixEngine/ClassTest.h
@@ -9,7 +9,7 @@ public:
 
 	void printClassInfo();
 private:
-	unsigned int mNumber;
+	size_t mNumber;
 	char* mName;
 };
